requirements.txt
pytest==3.7.2
Appium-Python-Client=0.49
selenium=3.141.0
